Ajustando algumas includes e detalhes de apresentação do programa

+++ decript.h
@@ -5,13 +5,13 @@
  * a string com a mensagem criptografada e o
  * inverso de e módulo (p - 1)*(q - 1). Não retorna nada.
  */
-void print_decrypted_text (char *str, long int d);
+void print_decrypted_text (char *str, int d);
 
 /*
  * Recebe o expoente "e" e "module" (= (p - 1)*(q - 1)).
  * Retorna o inverso de "e" módulo "module".
  */
-long int search_inverse_mod (long int e, long int module);
+int search_inverse_mod (int e, int module);
 
 /*
  * Recebe o valor do caractere criptografado "M" e o inver-
@@ -19,7 +19,7 @@ long int search_inverse_mod (long int e, long int module);
  * ractere decifrado m dado por M = m ^ (d * e) pela fór-
  * mula de Euler.
  */
-long int extract_root_mod (long int M, long int d);
+int extract_root_mod (int M, int d);
 
 /*
  * Chama as funções necessárias para decifrar a mensagem
